docs/2016_racak_martin/cd/elisa/elisa-server/requirements.txt
Django==1.11
django-cors-headers==2.0.2
django-filter==1.0.2
django-import-export==0.5.1
django-mptt==0.8.7
django-rest-swagger==2.1.2
djangorestframework-jwt==1.9.0
djangorestframework==3.6.2
drf-extensions==0.3.1
psycopg2==2.7.1
